new track version
Because of the GBA's 8bit audio DAC, I had to perform a lot of equalization to get the track to sound right. After listening to it a lot more and comparing to the original, I realized I wasn't doing it justice and started the equalization process from scratch.; sounds a lot closer to the original now. May have gone overboard on the bass, though...